+++ cmd_queue/base_queue.py
@@ -27,6 +27,48 @@ class Queue(ub.NiceRepr):
         self.all_depends = None
         self.named_jobs = {}
 
+    def change_backend(self, backend):
+        """
+        Create a new version of this queue with a different backend.
+
+        Currently metadata is not carried over. Submit an MR if you need this
+        functionality.
+
+        Example:
+            >>> from cmd_queue import Queue
+            >>> self = Queue.create(size=5, name='demo')
+            >>> self.submit('echo "Hello World"', name='job1a')
+            >>> self.submit('echo "Hello Revocable"', name='job1b')
+            >>> self.submit('echo "Hello Crushed"', depends=['job1a'], name='job2a')
+            >>> self.submit('echo "Hello Shadow"', depends=['job1b'], name='job2b')
+            >>> self.submit('echo "Hello Excavate"', depends=['job2a', 'job2b'], name='job3')
+            >>> self.submit('echo "Hello Barrette"', depends=[], name='jobX')
+            >>> self.submit('echo "Hello Overwrite"', depends=['jobX'], name='jobY')
+            >>> self.submit('echo "Hello Giblet"', depends=['jobY'], name='jobZ')
+            >>> serial_backend = self.change_backend('serial')
+            >>> tmux_backend = self.change_backend('tmux')
+            >>> slurm_backend = self.change_backend('slurm')
+            >>> airflow_backend = self.change_backend('airflow')
+            >>> serial_backend.rprint()
+            >>> tmux_backend.rprint()
+            >>> slurm_backend.rprint()
+            >>> airflow_backend.rprint()
+        """
+        new = Queue.create(backend=backend)
+        for job_name, job in self.named_jobs.items():
+            new_depends = []
+            if job.depends:
+                for dep in job.depends:
+                    new_dep = new.named_jobs[dep.name]
+                    new_depends.append(new_dep)
+            # TODO: carry over metadata
+            new.submit(job.command, depends=new_depends, name=job.name)
+        return new
+
+        for job in self.jobs:
+            new.submit(job.commands)
+            pass
+
     def __len__(self):
         return self.num_real_jobs
 
@@ -82,7 +124,13 @@ class Queue(ub.NiceRepr):
                         depends = [depends]
                     depends = self.all_depends + depends
                 kwargs['depends'] = depends
-            job = serial_queue.BashJob(command, **kwargs)
+            depends = kwargs.pop('depends', None)
+            if depends is not None:
+                # Resolve any strings to job objects
+                depends = [
+                    self.named_jobs[dep] if isinstance(dep, str) else dep
+                    for dep in depends]
+            job = serial_queue.BashJob(command, depends=depends, **kwargs)
         else:
             # Assume job is already a bash job
             job = command
@@ -112,17 +160,21 @@ class Queue(ub.NiceRepr):
 
     @classmethod
     def create(cls, backend='serial', **kwargs):
-        from cmd_queue import tmux_queue
-        from cmd_queue import serial_queue
-        from cmd_queue import slurm_queue
         if backend == 'serial':
+            from cmd_queue import serial_queue
             kwargs.pop('size', None)
             self = serial_queue.SerialQueue(**kwargs)
         elif backend == 'tmux':
+            from cmd_queue import tmux_queue
             self = tmux_queue.TMUXMultiQueue(**kwargs)
         elif backend == 'slurm':
+            from cmd_queue import slurm_queue
             kwargs.pop('size', None)
             self = slurm_queue.SlurmQueue(**kwargs)
+        elif backend == 'airflow':
+            from cmd_queue import airflow_queue
+            kwargs.pop('size', None)
+            self = airflow_queue.AirflowQueue(**kwargs)
         else:
             raise KeyError
         return self
@@ -154,8 +206,8 @@ class Queue(ub.NiceRepr):
         Builds a networkx dependency graph for the current jobs
 
         Example:
-            >>> from cmd_queue.tmux_queue import *  # NOQA
-            >>> self = TMUXMultiQueue(5, 'foo')
+            >>> from cmd_queue import Queue
+            >>> self = Queue.create(size=5, name='foo')
             >>> job1a = self.submit('echo hello && sleep 0.5')
             >>> job1b = self.submit('echo hello && sleep 0.5')
             >>> job2a = self.submit('echo hello && sleep 0.5', depends=[job1a])